[Rfc] New PlaceHolder management in order to insert instantiation place holders.

+++ asm-2016-05-26-valhalla-patch/backport-v0.1/resources/EqualatorMain.java
@@ -6,13 +6,13 @@ import java.util.AbstractMap;
 public class EqualatorMain {
 
     public static void main(String[] args) {
-        Holder<double> h = new Holder<double>(4.2);
-        Holder<double> h2 = new Holder<double>(4.2);
-        Equalator<double> eq = new Equalator<double>(h, h2);
+        Holder<float> h = new Holder<float>(4.2f);
+        Holder<float> h2 = new Holder<float>(4.2f);
+        Equalator<float> eq = new Equalator<float>(h, h2);
         System.out.println(Boolean.toString(eq.isEquals()));
         h.f();
 
-        QuadrupleHolder<double, float, byte, char> quadrupleHolder = new QuadrupleHolder<double, float, byte, char>(6.6, 2.2f, (byte)3, 'a');
+        QuadrupleHolder<float, float, byte, char> quadrupleHolder = new QuadrupleHolder<float, float, byte, char>(6.6f, 2.2f, (byte)3, 'a');
         quadrupleHolder.print();
     }
 }
